python/j0bz/requirements.txt
python-libnmap
termcolor